change: fetch関数を切り出し
